backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9

# Database - SQLAlchemy ORM with async support
sqlalchemy==2.0.29
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
passlib==1.7.4
bcrypt==4.1.2
python-jose==3.3.0
cryptography==42.0.5

# Environment variables management
python-dotenv==1.0.1

# CORS for React frontend integration
fastapi-cors

# Email functionality (for team invites, notifications)
fastapi-mail==1.4.1
jinja2==3.1.3

# File storage for note attachments
python-magic==0.4.27
aiofiles==23.2.1

# Background tasks & scheduling
apscheduler==3.10.4

# Validation and serialization
pydantic==2.6.4
pydantic-settings==2.2.1

# Utilities
python-dateutil==2.8.2
httpx==0.27.0
requests==2.31.0

# Pagination
fastapi-pagination==0.12.20

# Rate limiting (optional but recommended for production)
slowapi==0.1.8

# Testing dependencies (for when you start testing)
pytest==7.4.4
pytest-asyncio==0.23.5
httpx==0.27.0
pytest-cov==4.1.0

# Development tools
black==24.3.0
isort==5.13.2
flake8==7.0.0
mypy==1.9.0
pre-commit==3.7.0

# Additional utilities
tenacity==8.2.3  # For retry logic
python-slugify==8.0.4  # For URL-friendly slugs